gamers_mas/pyproject.toml
[project]
name = "game-deal-mas"
version = "0.1.0"
description = "Mock-data-only SPADE MAS prototype for demonstrating multi-agent coordination."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "spade==4.1.2",
  "slixmpp-multiplatform==1.12.0",
  "python-dotenv==1.2.2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
